+++ Online_Marketpace/src/by/fpmibsu/OnlineMarketplace/entity/Review.java
@@ -4,25 +4,25 @@ public class Review {
     public Review() {
     }
 
-    public Review(int ID_user_, int ID_product_, String review_) {
+    public Review(long ID_user_, long ID_product_, String review_) {
         this.ID_user_ = ID_user_;
         this.ID_product_ = ID_product_;
         this.review_ = review_;
     }
 
-    public int getID_user_() {
+    public long getID_user_() {
         return ID_user_;
     }
 
-    public void setID_user_(int ID_user_) {
+    public void setID_user_(long ID_user_) {
         this.ID_user_ = ID_user_;
     }
 
-    public int getID_product_() {
+    public long getID_product_() {
         return ID_product_;
     }
 
-    public void setID_product_(int ID_product_) {
+    public void setID_product_(long ID_product_) {
         this.ID_product_ = ID_product_;
     }
 
@@ -43,7 +43,7 @@ public class Review {
                 '}';
     }
 
-    private int ID_user_;
-    private int ID_product_;
+    private long ID_user_;
+    private long ID_product_;
     private String review_;
 }
